+++ src/Post/use-cases/delete-post-use-case.ts
@@ -1,10 +1,14 @@
 import { CommandHandler } from '@nestjs/cqrs';
 import { PostRepository } from '../postRepository';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 import { BlogRepository } from '../../Blog/blogRepository';
 
 export class DeletePostUseCaseCommand {
-  constructor(public postId: string, public blogId: string) {}
+  constructor(
+    public postId: string,
+    public blogId: string,
+    public userId: string,
+  ) {}
 }
 
 @CommandHandler(DeletePostUseCaseCommand)
@@ -21,12 +25,16 @@ export class DeletePostUseCase {
         'blogId not found blog /DeletePostUseCaseCommand',
       );
     }
+    if (blog.userId !== command.userId) {
+      throw new ForbiddenException('blog ne User / UpdateBlogUseCase/');
+    }
     const post = await this.postRepository.getPost(command.postId);
     if (!post) {
       throw new NotFoundException(
         'postId not found post /DeletePostUseCaseCommand',
       );
     }
+
     return this.postRepository.deletePost(command.postId);
   }
 }
